Add files via upload
A mobile app signup flow is the process of users creating an account and signing up for an
app. The goal of a good signup flow is to make it easy and quick for users to sign up, while
also collecting the necessary information from them
key requirement : 
Short and sweet : Users are more likely to complete a signup flow if it is short and easy to follow
Useful: The signup flow should collect the necessary information from users, but it should
not ask for too much information.
Secure: The signup flow should be secure, so that users' personal information is protected.